requirements/development.txt
-r base.txt
alabaster==0.7.6
Babel==2.1.1
defusedxml==0.4.1
django-allauth==0.23.0
docutils==0.12
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
psycopg2==2.6.1
Pygments==2.0.2
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
